+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- `User`, `EmailUser`, `EmailUserManager`
+
 ## [0.0.1] - 2026-02-26
 
 ### Added